style: 채팅 디자인 변경 + 공통 스크롤 관련 퍼블리싱
