+++ apps/client/src/actions/chatting/chattingAction.ts
@@ -7,8 +7,8 @@ import {
 } from '../../components/types/main/chatting/chattingTypes';
 import { commonResType } from '../../components/types/ResponseTypes';
 
-const userUuid = 'e782841e-7fcd-47c5-93e7-50203b3a0a99';
-const mentoringUuid = '3ec830b8-fade-4103-afef-4c633927c012';
+const userUuid = 'c120841a-7dd0-4967-a7a8-ed1daf2544d8';
+const mentoringUuid = 'f2a5b181-f1c3-4ad9-aa73-3d1bca4f5ad3';
 const mentoringSessionUuid = 'ac419217-cb98-4334-8b78-8126aa0e57aa';
 
 // 기존 채팅 데이터 불러오기
@@ -87,5 +87,63 @@ export async function getChatProfile({ userUuid }: { userUuid: string }) {
   } catch (error) {
     console.error('채팅 상대 프로필 조회 실패 : ', error);
     return { nickName: '', profileImageUrl: '' };
+  }
+}
+
+// 채팅방 입장
+export async function postEnterChat({
+  userUuid,
+  nickname,
+  // mentoringSessionUuid,
+}: {
+  userUuid: string;
+  nickname: string;
+  // mentoringSessionUuid: string;
+}) {
+  'use server';
+  try {
+    const res = await fetch(
+      `${process.env.CHATSERVICE_URL}/api/v1/chat/join/${mentoringSessionUuid}?nickName=${nickname}`,
+      {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'userUuid': userUuid,
+        },
+      }
+    );
+    return;
+  } catch (error) {
+    console.error('채팅방 입장 실패 : ', error);
+    return {};
+  }
+}
+
+// 채팅방 퇴장
+export async function postOutChat({
+  userUuid,
+  nickname,
+  // mentoringSessionUuid,
+}: {
+  userUuid: string;
+  nickname: string;
+  // mentoringSessionUuid: string;
+}) {
+  'use server';
+  try {
+    const res = await fetch(
+      `${process.env.CHATSERVICE_URL}/api/v1/chat/leave/${mentoringSessionUuid}?nickName=${nickname}`,
+      {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'userUuid': userUuid,
+        },
+      }
+    );
+    return;
+  } catch (error) {
+    console.error('채팅방 퇴장장 실패 : ', error);
+    return {};
   }
 }
