+++ apps/client/src/app/(main)/mentor/page.tsx
@@ -19,11 +19,12 @@ export default async function Page() {
           <span className="bg-slate-500 rounded-md py-1 px-2 text-white mr-2">
             안내
           </span>
-          현재 {}님의 멘토링은 {mentoringListData.length}개의 멘토링을 개설하고
-          있습니다.
+          현재 {}님의 멘토링은{' '}
+          {mentoringListData.length ? mentoringListData.length : 0}개의 멘토링을
+          개설하고 있습니다.
         </p>
         <Link
-          className="flex flex-row items-center bg-adaptorsGray text-md rounded-xl px-4 py-2 gap-x-2 text-white hover:bg-adaptorsBlue font-extrabold"
+          className="flex flex-row items-center bg-adaptorsGray text-md rounded-xl mr-2 px-4 py-2 gap-x-2 text-white hover:bg-adaptorsBlue font-extrabold"
           href={`/mentor/mentoring/add`}
         >
           <PlusIcon />
