+++ apps/client/src/app/(main)/mentor/schedule/page.tsx
@@ -14,7 +14,7 @@ export default async function Page() {
     <div className="w-full">
       <CalendarHeader />
       <div className="w-full h-full">
-        <div className="h-full pl-4 mb-6 bg-[#FAFAFE]">
+        <div className="h-full px-4 mb-6 bg-[#FAFAFE]">
           <CalendarContent scheduleList={scheduleList} />
         </div>
       </div>
